Header setup, Refactor: centralize shared data types and simplify Header/Footer structure
