+++ frontend-next-migration/src/shared/const/socialSectionMenu.tsx
