+++ frontend-next-migration/src/shared/types/auth.ts
