+++ frontend-next-migration/src/shared/types/types.ts
