+++ frontend-next-migration/src/shared/ui/v2/Footer/ui/FooterMobile/FooterMobile.test.tsx
@@ -1,5 +1,5 @@
 import { render, screen } from '@testing-library/react';
-import { SocialIconLink, Texts } from '@/widgets/Footer/model/types/types';
+import { SocialIconLink, Texts } from '@/shared/types/types';
 import FooterMobile from './FooterMobile';
 import { beforeEach } from 'node:test';
 import { useClientTranslation } from '@/shared/i18n';
