+++ frontend-next-migration/src/widgets/Footer/ui/FooterDesktop/FooterDesktop.stories.tsx
@@ -1,5 +1,5 @@
 import type { Meta, StoryObj } from '@storybook/nextjs';
-import { socialIconLinks } from '../../model/data/socialSectionMenu';
+import { socialIconLinks } from '../../../../shared/const/socialSectionMenu';
 import FooterDesktop from './FooterDesktop';
 
 const meta = {
