+++ frontend-next-migration/src/widgets/Footer/ui/FooterDesktop/FooterDesktop.tsx
@@ -1,6 +1,6 @@
 import { memo } from 'react';
 import { Container } from '@/shared/ui/Container';
-import { SocialIconLink, Texts } from '../../model/types/types';
+import { SocialIconLink, Texts } from '../../../../shared/types/types';
 import { Rights } from '../Rights/Rights';
 import { SocialSection } from '../SocialSection/SocialSection';
 import { Title } from '../Title/Title';
